Update README.md
The InNEED Online Academy project focuses on building a comprehensive and scalable database system to manage the core operations of an e-learning platform. InNEED Academy offers a wide range of online courses and educational programs to learners across various disciplines, supporting both personal and professional development. The platform is designed to be flexible and accessible, catering to a wide audience across the UK.

The goal of this project was to develop a robust MySQL database to streamline operations related to course offerings, student enrollment, payment processing, and certificate issuance. The database supports multiple user roles, each with its own set of functionalities, ensuring secure access and efficient data management.